Removed Frontend folder
Proceeding with merging

+++ Jenkinsfile
@@ -11,12 +11,23 @@ pipeline{
         stages{
             stage('Clone Repo'){
                 steps{
+                    script{
+                        if (env.rollback == 'false'){
+                        load "./ansible/.envvars/tf_ansible.groovy"
                         sh '''
                         rm -rf Final-Project
                         git clone https://github.com/makhdoomshabir/Final-Project.git
                         cd Final-Project
-                        sudo -E MYSQL_ROOT_PASSWORD=${env.MYSQL_ROOT_PASSWORD} DB_PASSWORD=${env.DB_PASSWORD} TEST_DATABASE_URI=${env.TEST_DATABASE_URI} SECRET_KEY=${env.SECRET_KEY} docker-compose build -d
+
+                        export MYSQL_DATABASE=test_db
+                        export MYSQL_ROOT_PASSWORD=root
+                        export MYSQL_USER=admin
+                        export MYSQL_PASSWORD=password
+
+                        sudo -E MYSQL_ROOT_PASSWORD=root DB_PASSWORD=root TEST_DATABASE_URI=mysql+pymysql://admin:password@test-db.c66nh3bppyv6.eu-west-2.rds.amazonaws.com:3306/testdb docker-compose build -d
                         '''
+                        }
+                    }
                 }
             }
             stage('Tag & Push Images'){
